+++ DesignToolbox/DesignToolbox/Pages/Components/Actions/Button/ButtonConfigurationView.swift
@@ -32,10 +32,6 @@ final class ButtonConfigurationModel: ComponentConfiguration {
     @Published var layout: ButtonLayout {
         didSet { updateCode() }
     }
-    
-    @Published var flipIcon: Bool {
-        didSet { updateCode() }
-    }
 
     @Published var flipIcon: Bool {
         didSet { updateCode() }