+++ CHANGELOG.md
@@ -8,14 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Library] Icon flip option for `text input` component trailing action (Orange-OpenSource/ouds-ios#1041)
 - [Library] Icon flip option for `button` component (Orange-OpenSource/ouds-ios#1124)
 - [DesignToolbox] Support of `watchOS`
 - [DesignToolbox] Support of `tvOS`
 
 ### Changed
 
-- [Library] Add flip option for icon on the trailing action of the text input (Orange-OpenSource/ouds-ios#1041)
-- [Library] Add flip option for button with icon in layout (Orange-OpenSource/ouds-ios#1124)
 - [Library] Update `text input` to display the label on two lines and in the middle when no placeholder and one line at top otherwise (Orange-OpenSource/ouds-ios#1086) 
 
 ### Removed